refactor: defining the final architecture
