+++ src/application/resources/CreateUserResponse.ts
@@ -1,0 +1,11 @@
+import User from '@domain/entities/User'
+
+class CreateUserResponse {
+  id: string;
+
+  constructor(user: User) {
+    this.id = user.id.toString()
+  }
+}
+
+export default CreateUserResponse
