+++ src/domain/entities/Authentication.ts
@@ -1,0 +1,16 @@
+import User from '@domain/entities/User'
+import { generateJwtToken } from '@utils/token'
+
+class Authentication {
+  user: User
+
+  constructor(user: User) {
+    this.user = user
+  }
+
+  get token(): string {
+    return generateJwtToken({ id: this.user.id })
+  }
+}
+
+export default Authentication
