+++ src/domain/exceptions/UserDuplicated.ts
