+++ src/domain/exceptions/UserNotFound.ts
