+++ src/domain/use_cases/updateUser.ts
@@ -1,0 +1,22 @@
+import UserNotFound from '@domain/exceptions/UserNotFound'
+import { updateUserById } from '@domain/repositories/UserRepository'
+import User from '@domain/entities/User'
+
+const updateUser = async (
+  id: string,
+  content: {
+    username?: string,
+    email?: string,
+    password?: string
+  }
+): Promise<User | null> => {
+  const updatedUser = await updateUserById(id, content)
+
+  if (updatedUser == null) {
+    throw new UserNotFound(`User not found for Id ${id}`)
+  }
+
+  return updatedUser
+}
+
+export default updateUser
