+++ src/infra/server/validate.ts
@@ -1,0 +1,16 @@
+import { validate } from 'class-validator'
+import ValidationException from '@application/exceptions/ValidationException'
+import ValidationErrorResponse from '@application/resources/ValidationErrorResponse'
+
+const validateRequest = async (
+  requestObj: object
+): Promise<void> => {
+  const errors = await validate(requestObj)
+
+  if (errors.length > 0) {
+    const response = new ValidationErrorResponse(errors)
+    throw new ValidationException(response)
+  }
+}
+
+export { validateRequest }
