--- src/resources/CreateUserResponse.ts
@@ -1,11 +1,0 @@
-import User from '@domain/User'
-
-class CreateUserResponse {
-  id: string;
-
-  constructor(user: User) {
-    this.id = user.id.toString()
-  }
-}
-
-export default CreateUserResponse
