--- src/use_cases/deleteUser.ts
@@ -1,14 +1,0 @@
-import UserNotFound from '@exceptions/UserNotFound'
-import { deleteUserById } from '@repositories/UserRepository'
-import { findUserById } from '@repositories/UserRepository'
-
-const deleteUser = async (id: string): Promise<void> => {
-  const user = await findUserById(id)
-  if (user == null) {
-    throw new UserNotFound(`User not found for Id ${id}`)
-  }
-
-  await deleteUserById(id)
-}
-
-export default deleteUser
